修改参数类型

+++ account-service/src/main/java/com/dili/account/dao/IUserCardDao.java
@@ -55,5 +55,5 @@ public interface IUserCardDao {
 	 * @param version
 	 * @return
 	 */
-	int updateState(@Param("accountId") Long accountId, @Param("state") Long state, @Param("version") Integer version);
+	int updateState(@Param("accountId") Long accountId, @Param("state") Integer state, @Param("version") Integer version);
 }
